Move ScoreProcessor and Score to Scoring/.
