+++ osu.Desktop.VisualTests/Tests/TestCaseLeaderboard.cs
@@ -4,9 +4,9 @@
 using OpenTK;
 using osu.Framework.Graphics;
 using osu.Framework.Screens.Testing;
-using osu.Game.Modes;
 using osu.Game.Modes.Mods;
 using osu.Game.Modes.Osu.Mods;
+using osu.Game.Modes.Scoring;
 using osu.Game.Screens.Select.Leaderboards;
 using osu.Game.Users;
