+++ osu.Game.Modes.Catch/Scoring/CatchScoreProcessor.cs
@@ -3,9 +3,10 @@
 
 using osu.Game.Modes.Catch.Judgements;
 using osu.Game.Modes.Catch.Objects;
+using osu.Game.Modes.Scoring;
 using osu.Game.Modes.UI;
 
-namespace osu.Game.Modes.Catch
+namespace osu.Game.Modes.Catch.Scoring
 {
     internal class CatchScoreProcessor : ScoreProcessor<CatchBaseHit, CatchJudgement>
     {
