+++ osu.Game.Modes.Mania/ManiaRuleset.cs
@@ -9,6 +9,8 @@ using osu.Game.Modes.Mods;
 using osu.Game.Modes.UI;
 using osu.Game.Screens.Play;
 using System.Collections.Generic;
+using osu.Game.Modes.Mania.Scoring;
+using osu.Game.Modes.Scoring;
 
 namespace osu.Game.Modes.Mania
 {
