+++ osu.Game.Modes.Mania/Scoring/ManiaScoreProcessor.cs
@@ -3,9 +3,10 @@
 
 using osu.Game.Modes.Mania.Judgements;
 using osu.Game.Modes.Mania.Objects;
+using osu.Game.Modes.Scoring;
 using osu.Game.Modes.UI;
 
-namespace osu.Game.Modes.Mania
+namespace osu.Game.Modes.Mania.Scoring
 {
     internal class ManiaScoreProcessor : ScoreProcessor<ManiaBaseHit, ManiaJudgement>
     {
