+++ osu.Game.Modes.Osu/Scoring/OsuScore.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-namespace osu.Game.Modes.Osu
+using osu.Game.Modes.Scoring;
+
+namespace osu.Game.Modes.Osu.Scoring
 {
     internal class OsuScore : Score
     {
