+++ osu.Game.Modes.Osu/osu.Game.Modes.Osu.csproj
@@ -72,8 +72,8 @@
     <Compile Include="OsuAutoReplay.cs" />
     <Compile Include="OsuDifficultyCalculator.cs" />
     <Compile Include="OsuKeyConversionInputManager.cs" />
-    <Compile Include="OsuScore.cs" />
-    <Compile Include="OsuScoreProcessor.cs" />
+    <Compile Include="Scoring\OsuScore.cs" />
+    <Compile Include="Scoring\OsuScoreProcessor.cs" />
     <Compile Include="UI\OsuHitRenderer.cs" />
     <Compile Include="UI\OsuPlayfield.cs" />
     <Compile Include="OsuRuleset.cs" />
