+++ osu.Game.Modes.Taiko/UI/TaikoHitRenderer.cs
@@ -3,9 +3,11 @@
 
 using osu.Game.Beatmaps;
 using osu.Game.Modes.Objects.Drawables;
+using osu.Game.Modes.Scoring;
 using osu.Game.Modes.Taiko.Beatmaps;
 using osu.Game.Modes.Taiko.Judgements;
 using osu.Game.Modes.Taiko.Objects;
+using osu.Game.Modes.Taiko.Scoring;
 using osu.Game.Modes.UI;
 
 namespace osu.Game.Modes.Taiko.UI
