+++ osu.Game.Modes.Taiko/osu.Game.Modes.Taiko.csproj
@@ -60,7 +60,7 @@
     <Compile Include="TaikoDifficultyCalculator.cs" />
     <Compile Include="Objects\TaikoHitObject.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TaikoScoreProcessor.cs" />
+    <Compile Include="Scoring\TaikoScoreProcessor.cs" />
     <Compile Include="UI\HitTarget.cs" />
     <Compile Include="UI\InputDrum.cs" />
     <Compile Include="UI\DrawableTaikoJudgement.cs" />
