+++ osu.Game/Modes/Scoring/ScoreProcessor.cs
@@ -1,15 +1,15 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using osu.Framework.Configuration;
 using System;
 using System.Collections.Generic;
-using osu.Game.Modes.Judgements;
-using osu.Game.Modes.UI;
-using osu.Game.Modes.Objects;
+using osu.Framework.Configuration;
 using osu.Game.Beatmaps;
+using osu.Game.Modes.Judgements;
+using osu.Game.Modes.Objects;
+using osu.Game.Modes.UI;
 
-namespace osu.Game.Modes
+namespace osu.Game.Modes.Scoring
 {
     public abstract class ScoreProcessor
     {
