+++ osu.Game/Screens/Ranking/Results.cs
@@ -6,7 +6,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Transforms;
 using osu.Game.Graphics.Sprites;
-using osu.Game.Modes;
+using osu.Game.Modes.Scoring;
 using osu.Game.Screens.Backgrounds;
 using OpenTK;
 using OpenTK.Graphics;
